Add files via upload
The Dining-Delight is online restaurant platform is a static website designed for food and drink ordering. It features a dynamic 3D menu that allows users to easily browse and order dishes and drinks. The site also provides information about the chefs, including reviews from reputable food critics, offering customers insights into the culinary expertise behind the offerings. Additionally, the website includes a feedback section where customers can leave their comments and reviews. To enhance user engagement, a "Contact Us" page is available, providing details for users to get in touch with the restaurant. The website is built using HTML for structure, CSS for styling, and JavaScript for interactive functionalities, creating an immersive and user-friendly dining experience